@@ -6,7 +6,7 @@ description: Demonstrates how to use the geocoding plugin.
 publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -23,8 +23,8 @@ dependencies:
 
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
-  cupertino_icons: ^0.1.3
-  url_launcher: ^5.4.11
+  cupertino_icons: ^1.0.2
+  url_launcher: ^6.0.2
 
 
 dev_dependencies:
