@@ -3,22 +3,22 @@ description: A common platform interface for the geocoding plugin.
 homepage: https://github.com/baseflow/flutter-geocoding/tree/master/geocoding_platform_interface
 # NOTE: We strongly prefer non-breaking changes, even at the expense of a
 # less-clean API. See https://flutter.dev/go/platform-interface-breaking-changes
-version: 1.0.1+1
+version: 2.0.0
 
 dependencies:
   flutter:
     sdk: flutter
 
-  meta: ^1.1.8
-  plugin_platform_interface: ^1.0.2
+  meta: ^1.3.0
+  plugin_platform_interface: ^2.0.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
 
-  effective_dart: ^1.2.1
-  mockito: ^4.1.1
+  effective_dart: ^1.3.0
+  mockito: ^5.0.0
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: '>=2.12.0 <3.0.0'
   flutter: ">=1.10.0"
