@@ -28,7 +28,7 @@ void main() {
 
   group('$MethodChannelGeocoding()', () {
     final log = <MethodCall>[];
-    MethodChannelGeocoding methodChannelgeocoding;
+    late MethodChannelGeocoding methodChannelgeocoding;
     var _mockCoordinatesNotFound = false;
 
     setUp(() async {
@@ -45,7 +45,6 @@ void main() {
             } else {
               return [_mockLocation.toJson()];
             }
-            break;
           case 'placemarkFromCoordinates':
             return [_mockPlacemark.toJson()];
           default:
@@ -65,7 +64,7 @@ void main() {
 
           // Act
           final locations =
-              await methodChannelgeocoding.locationFromAddress(address);
+              await (methodChannelgeocoding.locationFromAddress(address));
 
           // Assert
           expect(locations.length, 1);
@@ -113,10 +112,10 @@ void main() {
           final address = 'Gronausestraat, Enschede';
 
           // Act
-          final locations = await methodChannelgeocoding.locationFromAddress(
+          final locations = await (methodChannelgeocoding.locationFromAddress(
             address,
             localeIdentifier: 'nl-NL',
-          );
+          ));
 
           // Assert
           expect(locations.length, 1);
@@ -157,10 +156,10 @@ void main() {
 
           // Act
           final placemarks =
-              await methodChannelgeocoding.placemarkFromCoordinates(
+              await (methodChannelgeocoding.placemarkFromCoordinates(
             latitude,
             longitude,
-          );
+          ));
 
           expect(placemarks.length, 1);
           expect(placemarks.first, _mockPlacemark);
@@ -199,11 +198,11 @@ void main() {
 
           // Act
           final placemarks =
-              await methodChannelgeocoding.placemarkFromCoordinates(
+              await (methodChannelgeocoding.placemarkFromCoordinates(
             latitude,
             longitude,
             localeIdentifier: 'nl-NL',
-          );
+          ));
 
           expect(placemarks.length, 1);
           expect(placemarks.first, _mockPlacemark);
